requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
